refactor(app): moved modal to separate dir
